29 November 2023

+++ 11_Dec_2022/GPU_sph_Type_cooling/hfvGpuSphTyp_v4_cooling_debug_d2.cu
@@ -22,7 +22,7 @@ int main()
 
   float dt = 8e-7; //!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! This is only the first time step !!
 
-  const int Nngb_f = 64.0f; // used in smoothing func.
+  const float Nngb_f = 64.0f; // used in smoothing func.
   const int Nngb = 64;
   const int Ndown = Nngb - 10;
   const int Nup = Nngb + 10;
